First cut. only for windows now.
pass the path of file to addpymodules.py and it will add it to pydiction
dict.
Things need to be in mind are :
1. The modules should be importable in when the addpymodules runs.
2. This should be run by going to the directory in which complete-dict
is available.
3. this will create a bat file name complete-dict-modules.bat. run that
and it will add the modules.

My bat file to do all this is:
cd "C:\Program Files (x86)\Vim\vim73\ftplugin\pydiction"
python addpymodules.py c:\<projectname>\src
complete-dict-modules.bat